temporarily removed support for istream
Support for input streams requires bound and safety checking. However, I
have decided to drop support for input streams at the moment because it
would consume time for a feature that is not truly necessary.

I will definitely return to implementing support for input streams because
they would be helpful when reading from files.